requirements_dashboard.txt
nicegui==1.4.28
websocket-client==1.8.0
openpyxl==3.1.5
pandas==2.2.3

